+++ src/options/types/Address.ts
@@ -1,0 +1,6 @@
+import { IncognitoPattern } from "../enums/IncognitoPattern";
+
+export type Address = {
+  url: string;
+  type: IncognitoPattern;
+};