+++ webpack.config.js
@@ -3,6 +3,7 @@ const CopyWebpackPlugin = require('copy-webpack-plugin');
 const ChromeExtensionReloader = require('webpack-chrome-extension-reloader');
 const { VueLoaderPlugin } = require('vue-loader');
 const platform = require('platform');
+
 const { version } = require('./package.json');
 
 let mode;
@@ -14,34 +15,36 @@ const config = {
   mode,
   context: `${__dirname}/src`,
   entry: {
-    index: './index.js',
-    background: './background.js',
-    'popup/popup': './popup/popup.js',
-    'options/options': './options/options.js'
+    index: './index.ts',
+    background: './background.ts',
+    'popup/popup': './popup/popup.ts',
+    'options/options': './options/options.ts',
   },
   output: {
     path: `${__dirname}/dist`,
-    filename: '[name].js'
+    filename: '[name].js',
   },
   resolve: {
-    extensions: ['.js', '.vue']
+    extensions: ['.ts', '.js', '.vue'],
   },
   module: {
     rules: [
       {
         test: /\.vue$/,
-        loader: 'vue-loader'
+        loader: 'vue-loader',
       },
       {
-        test: /\.js$/,
-        loader: 'babel-loader',
-        exclude: /node_modules/
+        test: /\.ts$/,
+        loader: 'ts-loader',
+        options: {
+          appendTsSuffixTo: [/\.vue$/],
+        },
       },
       {
         test: /\.scss$/,
-        use: ['style-loader', 'css-loader', 'sass-loader']
-      }
-    ]
+        use: ['style-loader', 'css-loader', 'sass-loader'],
+      },
+    ],
   },
   plugins: [
     new VueLoaderPlugin(),
@@ -49,16 +52,16 @@ const config = {
       { from: 'icons', to: 'icons', ignore: ['icon.xcf', '.DS_Store'] },
       {
         from: 'popup/popup.html',
-        to: 'popup/popup.html'
+        to: 'popup/popup.html',
       },
       {
         from: 'options/options.html',
-        to: 'options/options.html'
+        to: 'options/options.html',
       },
       {
         from: 'manifest.json',
         to: 'manifest.json',
-        transform: content => {
+        transform: (content) => {
           const jsonContent = JSON.parse(content);
           jsonContent.version = version;
 
@@ -68,15 +71,15 @@ const config = {
           }
 
           return JSON.stringify(jsonContent, null, 2);
-        }
-      }
-    ])
-  ]
+        },
+      },
+    ]),
+  ],
 };
 
 if (config.mode === 'development' && platform.name === 'Chrome') {
   config.plugins = (config.plugins || []).concat([
-    new ChromeExtensionReloader()
+    new ChromeExtensionReloader(),
   ]);
 }